==> R/0_utilities.R <==
# Utilities
# Includes libraries, files/directories, and functions


# ---------------------------------------------------------------------
# Libraries
library(readr)
library(ggplot2)
library(tidyr)
library(dplyr)
library(purrr)
library(lubridate)
library(broom)
library(raster)
library(rgdal)
library(sf)


# ---------------------------------------------------------------------
# Files and Directories



# ---------------------------------------------------------------------
# Functions
